+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_container_functions.go
@@ -7,7 +7,6 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_value_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/port_spec_serializer"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/networking_sidecar"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/stacktrace"
@@ -135,31 +134,6 @@ func (backendCore *DockerKurtosisBackend) removeContainers(
 	return successfulContainers, erroredContainers
 }
 
-func (backendCore *DockerKurtosisBackend) getNetworkingSidecarContainersByEnclaveIdAndNetworkingSidecarGUIDs(
-	ctx context.Context,
-	enclaveId enclave.EnclaveID,
-	networkingSidecarGUIDs map[networking_sidecar.NetworkingSidecarGUID]bool,
-) (map[networking_sidecar.NetworkingSidecarGUID]*types.Container, error) {
-
-	enclaveContainers, err := backendCore.getEnclaveContainers(ctx, enclaveId)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting enclave status and containers for enclave with ID '%v'", enclaveId)
-	}
-
-	networkingSidecarContainers := map[networking_sidecar.NetworkingSidecarGUID]*types.Container{}
-	for _, container := range enclaveContainers {
-		if isNetworkingSidecarContainer(container) {
-			for networkingSidecarGuid := range networkingSidecarGUIDs {
-				userServiceGuid := service.ServiceGUID(networkingSidecarGuid)
-				if hasUserServiceGuidLabel(container, userServiceGuid){
-					networkingSidecarContainers[networkingSidecarGuid] = container
-				}
-			}
-		}
-	}
-	return networkingSidecarContainers, nil
-}
-
 func (backendCore *DockerKurtosisBackend) getUserServiceContainerByEnclaveIDAndUserServiceGUID(
 	ctx context.Context,
 	enclaveId enclave.EnclaveID,
@@ -217,14 +191,14 @@ func (backendCore *DockerKurtosisBackend) getUserServiceContainersByEnclaveIDAnd
 func getUserServiceContainerFromContainerListByEnclaveIdAndUserServiceGUID(
 	containers []*types.Container,
 	enclaveId enclave.EnclaveID,
-	userServiceGUID service.ServiceGUID) *types.Container {
+	userServiceGUID service.ServiceGUID) (*types.Container, error) {
 
 	for _, container := range containers {
-		if isUserServiceContainer(container) && hasEnclaveIdLabelAndUserServiceGuidLabel(container, enclaveId, userServiceGUID) {
-			return container
+		if isUserServiceContainer(container) && hasEnclaveIdLabel(container, enclaveId) && hasUserServiceGuidLabel(container, userServiceGUID) {
+			return container, nil
 		}
 	}
-	return nil
+	return nil, stacktrace.NewError("No user service container with user service GUID '%v' was found in container list '%+v'", userServiceGUID, containers)
 }
 
 func getServiceIdFromContainer(container *types.Container) (service.ServiceID, error) {
@@ -246,7 +220,7 @@ func isUserServiceContainer(container *types.Container) bool {
 	containerTypeValue, found := labels[label_key_consts.ContainerTypeLabelKey.GetString()]
 	if !found {
 		//TODO Do all containers should have container type label key??? we should return and error here if this answer is yes??
-		logrus.Debugf("Container with ID '%v' haven't label '%v'", container.GetId(), label_key_consts.ContainerTypeLabelKey.GetString())
+		logrus.Debugf("Container with ID '%v' does not have label '%v'", container.GetId(), label_key_consts.ContainerTypeLabelKey.GetString())
 		return false
 	}
 	if containerTypeValue == label_value_consts.UserServiceContainerTypeLabelValue.GetString() {
@@ -255,50 +229,7 @@ func isUserServiceContainer(container *types.Container) bool {
 	return false
 }
 
-func isNetworkingSidecarContainer(container *types.Container) bool {
-	labels := container.GetLabels()
-	containerTypeValue, found := labels[label_key_consts.ContainerTypeLabelKey.GetString()]
-	if !found {
-		//TODO Do all containers should have container type label key??? we should return and error here if this answer is yes??
-		logrus.Debugf("Container with ID '%v' haven't label '%v'", container.GetId(), label_key_consts.ContainerTypeLabelKey.GetString())
-		return false
-	}
-	if containerTypeValue == label_value_consts.NetworkingSidecarContainerTypeLabelValue.GetString() {
-		return true
-	}
-	return false
-}
 
-func hasEnclaveIdLabelAndUserServiceGuidLabel(
-		container *types.Container,
-		enclaveId enclave.EnclaveID,
-		userServiceGuid service.ServiceGUID) bool {
-
-	labels := container.GetLabels()
-	enclaveIdLabelValue, found := labels[label_key_consts.EnclaveIDLabelKey.GetString()]
-	if !found {
-		//TODO Do all containers should have enclave ID label key??? we should return and error here if this answer is yes??
-		logrus.Debugf("Container with ID '%v' haven't label '%v'", container.GetId(), label_key_consts.EnclaveIDLabelKey.GetString())
-		return false
-	}
-	if enclaveIdLabelValue == string(enclaveId) && hasUserServiceGuidLabel(container, userServiceGuid) {
-		return true
-	}
-	return false
-}
-
-func hasUserServiceGuidLabel(container *types.Container, userServiceGuid service.ServiceGUID) bool {
-	labels := container.GetLabels()
-	userServiceGuidLabelValueStr, found := labels[label_key_consts.EnclaveIDLabelKey.GetString()]
-	if !found {
-		return false
-	}
-	userServiceGuidLabelValue := service.ServiceGUID(userServiceGuidLabelValueStr)
-	if userServiceGuidLabelValue == userServiceGuid {
-		return true
-	}
-	return false
-}
 
 func getPrivatePortsFromContainerLabels(containerLabels map[string]string) (map[string]*port_spec.PortSpec, error) {
 	serializedPortSpecs, found := containerLabels[label_key_consts.PortSpecsLabelKey.GetString()]