+++ container-engine-lib/lib/backend_interface/objects/networking_sidecar/networking_sidecar.go
@@ -2,24 +2,22 @@ package networking_sidecar
 
 import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"net"
 )
 
-type NetworkingSidecarGUID string
-
 type NetworkingSidecar struct {
-	guid NetworkingSidecarGUID
-	privateIpAddr net.IP
+	serviceGuid service.ServiceGUID
+	privateIpAddr net.IP // TODO Delete this when we solve the static IP problem
 	enclaveId enclave.EnclaveID
 }
 
-func NewNetworkingSidecar(guid NetworkingSidecarGUID, privateIpAddr net.IP, enclaveId enclave.EnclaveID) *NetworkingSidecar {
-	return &NetworkingSidecar{guid: guid, privateIpAddr: privateIpAddr, enclaveId: enclaveId}
+func NewNetworkingSidecar(serviceGuid service.ServiceGUID, privateIpAddr net.IP, enclaveId enclave.EnclaveID) *NetworkingSidecar {
+	return &NetworkingSidecar{serviceGuid: serviceGuid, privateIpAddr: privateIpAddr, enclaveId: enclaveId}
 }
 
-
-func (sidecar *NetworkingSidecar) GetGuid() NetworkingSidecarGUID {
-	return sidecar.guid
+func (sidecar *NetworkingSidecar) GetServiceGuid() service.ServiceGUID {
+	return sidecar.serviceGuid
 }
 
 func (sidecar *NetworkingSidecar) GetPrivateIpAddr() net.IP {