+++ container-engine-lib/lib/backend_interface/objects/networking_sidecar/networking_sidecar_filters.go
@@ -1,7 +1,9 @@
 package networking_sidecar
 
+import "github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
+
 type NetworkingSidecarFilters struct {
-	// Disjunctive set of networking sidecar GUIDs to find networking sidecar for
+	// Disjunctive set of user service GUIDs to find networking sidecar for
 	// If nil or empty, will match all IDs
-	GUIDs map[NetworkingSidecarGUID]bool
+	GUIDs map[service.ServiceGUID]bool
 }